+++ bench.py
@@ -2,12 +2,12 @@ from json import dumps
 
 import pyperf
 
-from reparsec.json import parse
-from reparsec.json_scannerless import parse as sl_parse
+from reparsec.parsers.json import loads
+from reparsec.parsers.json_scannerless import loads as sl_loads
 
 DATA = dumps({"key_" + str(n): list(range(100)) for n in range(1000)})
 
 
 runner = pyperf.Runner()
-runner.bench_func("json_parser", lambda: parse(DATA))
-runner.bench_func("json_sl_parser", lambda: sl_parse(DATA))
+runner.bench_func("json_parser", lambda: loads(DATA))
+runner.bench_func("json_sl_parser", lambda: sl_loads(DATA))
