+++ src/reparsec/core/result.py
