+++ src/reparsec/core/types.py
@@ -6,27 +6,11 @@ from typing_extensions import Literal
 from .result import Result
 from .state import Ctx
 
-S = TypeVar("S")
 S_contra = TypeVar("S_contra", contravariant=True)
-V = TypeVar("V")
 V_co = TypeVar("V_co", covariant=True)
 
 
 RecoveryMode = Literal[None, False, True]
-
-
-def disallow_recovery(rm: RecoveryMode) -> RecoveryMode:
-    if rm is None:
-        return None
-    return False
-
-
-def maybe_allow_recovery(rm: RecoveryMode, r: Result[V, S]) -> RecoveryMode:
-    if rm is not None and r.consumed:
-        return True
-    return rm
-
-
 ParseFn = Callable[
     [S_contra, int, Ctx[S_contra], RecoveryMode],
     Result[V_co, S_contra]
