+++ src/reparsec/parsers/yamlish.py
@@ -1,6 +1,9 @@
 from typing import Tuple
 
-from .parser import Delay, Parser, alt, eof, indented, regexp, same, sl_run
+from reparsec.layout import indented, same
+from reparsec.parser import Delay, Parser, alt
+from reparsec.scannerless import regexp, run
+from reparsec.sequence import eof
 
 eol = regexp(r"\n\s*")
 ows = regexp(r"\s*")
@@ -15,5 +18,5 @@ pair.define((same(ident) << regexp("[ \t]*:[ \t]*")) + value)
 parser = ows >> pair.many().fmap(lambda kvs: dict(kvs)) << eof()
 
 
-def parse(source: str) -> object:
-    return sl_run(parser, source).unwrap()
+def loads(source: str) -> object:
+    return run(parser, source).unwrap()
