+++ tests/test_json_scannerless.py
@@ -2,9 +2,9 @@ from typing import List, Tuple
 
 import pytest
 
-from reparsec import json_scannerless
 from reparsec.output import ParseError
-from reparsec.parser import sl_run
+from reparsec.parsers import json_scannerless
+from reparsec.scannerless import run
 
 DATA_POSITIVE: List[Tuple[str, object]] = [
     (r"1", 1),
@@ -28,7 +28,7 @@ DATA_POSITIVE: List[Tuple[str, object]] = [
 
 @pytest.mark.parametrize("data, expected", DATA_POSITIVE)
 def test_positive(data: str, expected: object) -> None:
-    assert json_scannerless.parse(data) == expected
+    assert json_scannerless.loads(data) == expected
 
 
 DATA_NEGATIVE = [
@@ -48,7 +48,7 @@ DATA_NEGATIVE = [
 @pytest.mark.parametrize("data, expected", DATA_NEGATIVE)
 def test_negative(data: str, expected: str) -> None:
     with pytest.raises(ParseError) as err:
-        json_scannerless.parse(data)
+        json_scannerless.loads(data)
     assert str(err.value) == expected
 
 
@@ -91,7 +91,7 @@ DATA_RECOVERY: List[Tuple[str, object, str]] = [
 
 @pytest.mark.parametrize("data, value, expected", DATA_RECOVERY)
 def test_recovery(data: str, value: str, expected: str) -> None:
-    r = sl_run(json_scannerless.json, data, recover=True)
+    r = run(json_scannerless.parser, data, recover=True)
     assert r.unwrap(recover=True) == value
     with pytest.raises(ParseError) as err:
         r.unwrap()
