+++ tox.ini
@@ -29,6 +29,7 @@ deps =
 commands =
     coverage run -m pytest {posargs}
     coverage report
+    coverage html
     coverage xml
 deps =
     pytest
@@ -49,6 +50,11 @@ testpaths =
 [coverage:run]
 source = reparsec
 branch = true
+
+[coverage:report]
+exclude_lines =
+    pragma: no cover
+    @abstractmethod
 
 [isort]
 line_length = 79
